Skip sending empty sessions data payload (#98)
Currently even if there has been no session data generated we still send a payload to Bugsnag. If no session data has been generated then we should skip sending the payload

The most substantial change was in the test code. We needed to support providing a timeout to our test server when we are expecting a certain number of requests. This was added so that we could test the scenario above.

The functional change made was to check that session data exists before generating the payload to send.